+++ CHANGELOG.markdown
@@ -1,3 +1,8 @@
+### 0.7.1 / 
+* Modified nextval/currval syntax in JDBCHelper::SequenceWrapper
+ * PostgreSQL: nextval('[sequence]')
+ * Others: [sequence].nextval
+
 ### 0.7.0 / 2012/01/30
 * Helper methods for generating where clauses for TableWarppers
  * `JDBCHelper::SQL.expr` (`JDBCHelper::SQL` deprecated in favor of `expr`)